End to End code placed in for Web site - before testing

+++ src/ScheduMail.Core/RepositoryInterfaces/IWebSiteRepository.cs
@@ -5,7 +5,7 @@ namespace ScheduMail.Core.RepositoryInterfaces
     /// <summary>
     /// IWeb site repository to define web site ad store web site template details
     /// </summary>
-    public interface IWebSiteRepository : ICrudRepository<WebSite, int>
+    public interface IWebSiteRepository : ICrudRepository<WebSite, long>
     {
     }
 }
